fix: update expense date validation to allow optional string format
The `expenses.expensedDate` validation in the `monthlyBudgetRoutes.js` file has been updated. It now uses `optional().isString()` to allow the date to be an optional string, instead of a required Date object. This change accommodates cases where the expense date might not always be available or provided as a string.